requirements.txt
ccxt
pandas
numpy
matplotlib
python-dotenv
transformers
torch
scikit-learn
